Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> I agree with that Peter's simplification would be a good implementation enhancement too

It is already in the patch and a part of the original proposal (I guess Peter was just talking about adding it to the spec if adopted). However, we postponed the decision at the last conference call, so we cannot merge it yet.

Patch-set: 3
Reviewer: Gerrit User 6720 <6720@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>