Update patch set 11
Patch Set 11:

(3 comments)

> Patch Set 11:
> 
> (1 comment)
> 
> I'll let you take a first pass through this and see what you think, Andrey.
> 
> Any concerns about moving some of writing logic into grain_synthesis? It allows the file representation to use the bitstream specification.
> 
> Stand-alone app that takes in denoised and original will come later:
> https://aomedia-review.googlesource.com/c/aom/+/51684/1/examples/noise_model.c

Yes, there is a concern that it is not easy to check whether there was a normative change (at least for the decoder part). Also other table for syntax reading / writing are located where they used to be. However, maybe it is still OK to have the syntax in the film grain if it helps us to avoid code duplication.

Patch-set: 11
Reviewer: Gerrit User 6720 <6720@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>