Update patch set 2
Patch Set 2:

Irrespective of whether we revert https://aomedia-review.googlesource.com/c/aom/+/51921 or fix it, it will be good to disable this test temporarily until we have resolved all the issues. There have been other issues reported with this test as well.
Right now we are blocked by this failure and it has not been possible to get a clean build with any revert.

Patch-set: 2