Update patch set 6
Patch Set 6:

While looking at some other patches, I discovered that this commit causes a normative change.

Before this commit, cdf_num would be set to 0 whenever we're on the top or left edge of the frame, due to a special case in pick_spatial_seg_cdf. But that special case was removed in this commit, and so cdf_num is set to either 1 (if we're at the top or left edge, but not the top-left corner) or 2 (in the very top-left corner of the frame).

I have no preference either way on this change, but want to check that it is intentional.

Patch-set: 6
CC: Gerrit User 5975 <5975@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>