Update patch set 4
Patch Set 4:

Is this intended to be a normative change?

I think it causes a normative change when we find a valid motion vector in two different source reference frames.

For example, suppose both LAST_FRAME and BWDREF_FRAME both generate a motion vector in the stack.

In the previous code, it would end up picking the first valid entry in the stack (e.g. the motion vector for LAST_FRAME).  In the new code, it would end up picking the last valid entry in the stack (e.g. the motion vector for BWDREF_FRAME) because each valid entry overwrites the entries found before.

To be non-normative I think you could first work out which frames you want to project in forward order, and then visit them in reverse order.

Patch-set: 4
CC: Gerrit User 5189 <5189@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
CC: Gerrit User 5260 <5260@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
CC: Gerrit User 6145 <6145@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>