Update patch set 14
Patch Set 14:

(4 comments)

> Patch Set 11:
> 
> (3 comments)
> 
> > Patch Set 11:
> > 
> > (1 comment)
> > 
> > I'll let you take a first pass through this and see what you think, Andrey.
> > 
> > Any concerns about moving some of writing logic into grain_synthesis? It allows the file representation to use the bitstream specification.
> > 
> > Stand-alone app that takes in denoised and original will come later:
> > https://aomedia-review.googlesource.com/c/aom/+/51684/1/examples/noise_model.c
> 
> Yes, there is a concern that it is not easy to check whether there was a normative change (at least for the decoder part). Also other table for syntax reading / writing are located where they used to be. However, maybe it is still OK to have the syntax in the film grain if it helps us to avoid code duplication.

I was also a bit hesitant for that reason, and I didn't want to pull in too many dependencies on encoder structures (if trying to call the routines directly).

I've rewritten the file format to be ascii-based. In that way, there is no confusion between binary representations, and no need to share the code. It also makes it easier to visualize and edit parameters (current representation favors readability/parse-ability). The compromise is a slightly bigger file. But even if there were different grain parameters per frame (which there shouldn't be), for an hour long 30fps file it should only be on the order of 10MB or so.

Patch-set: 14
Reviewer: Gerrit User 8612 <8612@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>