Update patch set 6
Patch Set 6:

My understanding is that we need a way to tell the reference decoder which operating point to decode.

Also, my understanding is that a decoder should discard all OBUs that are not within the specified operating point because this affects the normative decoding process - especially when the bitstream contains two independent spatial resolutions each with their own keyframe.

Patch-set: 6
CC: Gerrit User 5260 <5260@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>