Update patch set 4
Patch Set 4:

> This is a non-normative change. In the example where both last and bwd frames generate valid mv, the mv from last frame will be stored in stack index ref_stamp = (MFMV_STACK_SIZE - 1). This ref_stamp index would reduce every time one valid reference frame is scanned. Hence the mv from bwd frame would be stored at a lower index value. Hence the mv from bwd would be picked in both cases.

Good point, I had missed that the process for searching the stack worked in increasing order, while the stack was filled in reverse order.

It does seem a little odd that we seem to give more priority to distant frames than we do to LAST_FRAME.  I wonder if doing this in the opposite order would give better compression?

Patch-set: 4