[NORMATIVE] Fix ref_frame_side setup
Patch 1609b94f disabled the ref_frame_side setup code when in
error-resilient mode, as at that point we did not explicitly
signal order hints, and so could not determine the correct value
of ref_frame_side for any dropped frames.

However, we now always signal the order hints of the reference
frames when in error-resilient mode. Thus we can work out the correct
values of ref_frame_side, even if frames have been dropped.
Since we are allowed to use MFMV in error-resilient mode, we now
want to set up the ref_frame_side array properly.

Note: We can keep the code which disables ref_frame_side setup when
!use_order_hint. That is purely an optimization - when !use_order_hint,
MFMV is disabled and so we won't use ref_frame_side anyway.

BUG=aomedia:1634

Change-Id: I09b4993e95ee1e1d627f90856a14f6138a9d39ad