Update patch set 1
Patch Set 1:

As I was doing the patch, I realized that this patch may not actually be required.
I will let Yue/Cheng verify this.

If we do check this in, it will be a normative change on the decoder side, as far as I can see.

Patch-set: 1
Reviewer: Gerrit User 5189 <5189@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>