[NORMATIVE] Fix issue in CfL Lossless Prediction
Normative change: In lossless mode (ONLY) the
DC_PRED required by CfL is computed at the
transform block level.

*This is currently broken and must be fixed*

We believe this fix is aligned with the "spirit"
of the spec with regards to lossless. This semantic
change is required, since lossless mode forces
transform size to 4x4.

Non-Normative change: The following changes are
required in libaom to resolve this issue.
However, none of these changes have alter the
bitstream or its semantic.

  * Disable DC_PRED caching
  * Add offset to subtract average
  * Add offset to prediction

BUG=aomedia:1212

Change-Id: Ib57ba18f6dc23c4ad6d31151d2473d2d8bb5d0c1