Update patch set 8
Patch Set 8:

In HEVC there is an analogous control called HighestTID.

This is an input to the normative decoding path and specifies which temporal layers to decode.

I had assumed we would also allow the desired operating point to be specified in the same way.

It sounds like this might be suitable for discussion in the working group as I think it will significantly affect conformance testing.

Specifying the operating point via which level is supported doesn't quite seem right because we can have 32 operating points, but only 16 levels, so it would be impossible to trigger half of the operating points.

Patch-set: 8