Update patch set 9
Patch Set 9:

> Patch Set 9: Code-Review+1
> 
> Looks good to me.  
> 
> I agree that adding the choice of operating point to decode should be in a separate patch as I think it may need wider discussion before a consensus is reached on the preferred approach.
> 
> (Probably not worth fixing, but there is a small type in initial_dispaly_delay)

We should fix the typo, but that can be a follow-up patch if it is easier for Soo-Chul. I am leaving that to him to decide. We shall get this and typo fix merged before end of today hopefully.

Patch-set: 9
Reviewer: Gerrit User 5005 <5005@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>