Set single_tile_decoding mode correctly
In the large scale tile coding, single_tile_decoding mode is set based
on whether or not loopfilter, cdef, and restoration are disabled. Move
single_tile_decoding setting down after the above filter information
is known, so that the correct mode is set while TILE_INFO_FIRST is
turned on.

Change-Id: I9cadeaacd525753072f1292c5c3c83014992e841