[NORMATIVE] Seq level flag to control mfmv usage
This flag allows usage of ref frame mvs to be disabled for the
entire sequence providing a type of error resilience where
non-E frames after a E frame can be parsed irrespective of
the encoding strategy for frames following an E-frame if
all references are at the same resolution.
If the flag is disabled at the frame level, then the
use_ref_frame_mvs flag is not sent and instead assumed to
be 0.
A similar flag for disabling warped motion can be added next,
to remove the restriction on resolution.

BUG=aomedia:1655

Change-Id: I694b7af3517972a8c235ded649dde36d81652def