Add sequence level flag to enable/disable ref mvs
This flag allows the ref frame mvs to be disabled for the
entire sequence providing a type of error resilience where
non-E frames after a E frame can be parsed irrespective of
the encoding strategy for frames following an E-frame.
If the flag is disabled at the frame level, then the
use_ref_frame_mvs flag is not sent and instead assumed to
be 0.

BUG=aomedia:1655

Change-Id: I694b7af3517972a8c235ded649dde36d81652def