Update patch set 4
Patch Set 4: Code-Review-1

This is changing the function get_segment_id() which is used by the decoder via av1_get_spatial_seg_pred(). However, I don't think a decoder change is required to fix this issue. I think there is an underlying issue in the encoder that this patch is papering over. Therefore I don't think we should make normative changes to the decoder to address it. I don't think this actually has any effect on the decoding process, but it is probably better to err on the side of caution at this stage.

Adding the assert() to av1_neg_interleave() is still a good idea, though.

Patch-set: 4
Reviewer: Gerrit User 5075 <5075@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
Label: Code-Review=-1