Update patch set 3
Patch Set 3: Code-Review+1

This looks like a fantastic idea.

Is someone writing the spec changes, or do you want me to have a go?

I think there are several small changes updates needed to the spec to make the bitstream conformance requirements accurate.  Some changes due to this change, but also more changes due to previous changes to the encoder which implicitly changed the conformance requirements (but I missed as they did not change the decoder implementation).

In particular, we need to change the read bit process within the symbol decoder to return 0's if we have gone past the maximum bits allowed, and to change the exit symbol somehow to check the location of the trailing bit, and relax the requirement of not reading more bits than are available.

For example, at the moment the init_symbol process reads 15 bits, but this needs to be modified for cases when the entire tile has fewer than 15 bits available.

Patch-set: 3
Reviewer: Gerrit User 5260 <5260@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
Label: Code-Review=+1