Preserve Lossless mode after in update_state
When the update_state function copies the segment id
from the segment map, it can change whether an mbmi
is lossy or lossless. This update occurs after the
mode decision as been taken.

To avoid getting into an illegal state, and to better
match the mode decision parameters, we preserve the
lossless/lossy state of the mbmi during update_state
function.

BUG=aomedia:1665

Change-Id: Id03574f6f9aba8e433d246f5d49fada7803bedbb