Speed feature: reuse ref0 chosen by counterpart partition
When going through different partitions, some mbs could have been
fully tested in a previously tested partition type. For example,
block 2 in horzA has been tested as block 1 in horz except that
the above context is different.

This patch links those blocks to their counterparts by setting up
pointers to the references in PICK_MODE_CTX. The framework could
be extended later to consider parent blocks as well.

Built on top of the framework, in RDO, when a valid reference is
available, ref_frame[0] is reused so we skip testing other first
frames.

It is tentatively added as a speed 1 feature.

Performance degradation: -0.103% (lowres, speed 1, 30 frames)
Speed up: 7.2% (soccer, 30 frames)

Change-Id: I113b16e2b9cdd77e7b1ffd0c9d0bf741e0464d86