Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I do not think it is 'fully' adopted yet.

I was relying on the assumption that to be enabled by default, an experiment has to be fully adopted. Is it generally the case, and there's an exception for this specific experiment?

Just for my information: is there a reasonable chance that we go back to the old (currently-non default) behavior (which involves random number generation inside the decoder)?

Patch-set: 1
Reviewer: Gerrit User 6000 <6000@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>