Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > I do not think it is 'fully' adopted yet.
> 
> I was relying on the assumption that to be enabled by default, an experiment has to be fully adopted. Is it generally the case, and there's an exception for this specific experiment?
> 
> Just for my information: is there a reasonable chance that we go back to the old (currently-non default) behavior (which involves random number generation inside the decoder)?

When 'provisionally adopted' we turn the tool on by default but the config macro stays. When the tool clears TAPAS we call it 'fully adopted' and remove the macro. This is what we have done in most cases.

Patch-set: 1
Reviewer: Gerrit User 5189 <5189@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>