[CFL] AVX2 Version of luma_subsampling_420_hbd
Based on the observation that for smaller widths (4xN
and 8xN), the AVX2 code tends to be slower than its
SSSE3 counterpart.

Because of the laning in _mm256_hadd_epi64, an extra
_mm256_permute4x64_epi64 operation is required for
AVX2 when compared to SSSE3. In light of this, blocks
of width 16, are slower in AVX2 than in SSSE3.

The AVX2 code now calls the SSSE3 functions when
width < 32.

We include unit tests for conformance and speed.

AVX2/CFLSubsampleHBDSpeedTest (i7-7820X)
  4x4: C time = 96 us, SIMD time = 56 us (~1.7x)
  8x8: C time = 319 us, SIMD time = 96 us (~3.3x)
  16x16: C time = 1632 us, SIMD time = 243 us (~6.7x)
  32x32: C time = 7973 us, SIMD time = 769 us (~10x)

Change-Id: I3234668060908018b18f37f045011d6f9751ff81