Update patch set 4
Patch Set 4:

> Patch Set 4:
> Inputs to inverse transforms should be 20 bits wide in the worst case because of normative clamping operations we have defined.
> The only case we should worry about here is where a 20-bit input x is multiplied by NewSqrt2=5793=pow(2,12.5), as x*NewSqrt2 may overflow the 32-bit range.
> Instead of computing x*NewSqrt2>>12, a decoder may choose to implement it as x+(x*(NewSqrt2-4096)>>12) to avoid possible 32-bit range overflow.
> A decoder may also choose to clamp x to a tighter range by relying on the knowledge that the result of the inverse transform will be further clamped.

Thanks for your insights, this makes a lot of sense to me.

What I was having a problem with, initially, was the source code comment. My first patchset in this CL was actually about simply removing it.

I don't agree with the interpretation that "if the data exceeds 32-bit, then the input bitstream is invalid/corrupted" (see previous comments in this CL).

Do we agree that this is actually not the case, and that the cast(int64_t) is actually necessary (although, there are indeed some other ways to achieve the same calculation that don't require 64 bit intermediates).

My understanding is that by simply removing the cast, one wouldn't be able to decode some *valid* bitstreams. Do we agree on this?

Please me know what you think.

Patch-set: 4
Reviewer: Gerrit User 6000 <6000@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>