Reset wm_params[].invalid properly
When an invalid local warped model is found, invalid will be set
to 1. If not resetting it, the next WM block will inherit this
value by mistake.
Note: libaom encoder will not transmit an invalid model therefore
at the decoder side invalid will not be 1. But better to guard the
decoding logic properly.

BUG=aomedia:1699

Change-Id: I51aed86d8ed00440a484d45dbf088fd0ab52d61d