Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> > Patch Set 5:
> > I agree that the comment is incorrect.
> > A valid bitstream is any sequence of bits that is not deemed nonconforming by the spec.
> > For strict conformance intermediate 64-bit values are needed with the current implementation and current spec (AFAIK).
> > Encoders probably don't have to worry about these cases but strictly compliant decoders have to.
> 
> This matches my understanding.
> Hui, if you agree with this, could you please remove your -1 ?
> The comment I'm trying to remove here basically represents an implicit range check (aka 'normative range'), which is confusing.

For identity transform(basically just multiply the prediction residue by 8 or 4), the normal input signal is at most 16 bits, so it won't overflow 32 bits after multiplication with NewSqrt2. It overflows only when the input signal is corrupted, which means the bitstream is non-conformant.
Here is the CL in which the comment was added:
https://aomedia-review.googlesource.com/c/aom/+/50844

Patch-set: 6
Reviewer: Gerrit User 5187 <5187@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>