[NOT-normative] Fix CfL prediction for semi-lossless frames
When some segment IDs use lossless mode and others do not,
it is possible for the luma pixel buffer used by CfL to be
incorrectly updated, leading to bad predictions. As such
lossless mode must not be considered when storing.

Fix ignore lossless/lossy when storing luma pixels. This
stores more pixels than necessary, but that can be
addressed in future non-normative changes.

BUG=aomedia:1707

Change-Id: Ib71d2811d4211f5ea9236ba685a93f0f1ff08ec1