Fix early terminate in jnt_comp search
In RDO, now we search jnt_comp first, and then other compound modes.

Whenever we encounter invalid rd, e.g., mv out of bounds, we set
early termination flag and continue to the next search. It has a
potential bug that when the first search is valid, but the second
is not. The encoder will end up with returning invalid rd, due to
early termination is found in the second search.

This CL let handle_inter_mode return correct rd.

STATS_CHANGE
Performance: xx%
Speed: xx%

Change-Id: Id2d8690ff9ca0a13a2cbcba6fc2ee86424065e13