Update patch set 7
Patch Set 7:

> Patch Set 6:
> 
> > Patch Set 6:
> > 
> > > Patch Set 6: Code-Review+1
> > 
> > Does this mean you're happy for this to be merged now? If so, Deb, please could you +2 this and merge?
> 
> If you're happy, I'm happy... This patch is well crafted. I would prefer a solution that resovle the underlining issue of semi-lossless RDO. However, given the timeframe this fix does not reduce coding performance so I don't have an issue

The issue wasn't with the encoder per se - indeed, the stream I saw a problem with wasn't generated by the reference encoder at all, but by our internal tools. So there's no RDO issue to fix.

Instead, the issue was that the common code was doing something which isn't quite right, in a situation which is allowed by the bitstream format (even if it is a bit of a weird situation).

Patch-set: 7
Reviewer: Gerrit User 5975 <5975@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>