Add ability to drop invisible frames from a test
This allows us to drop invisible frames before an S or E frame.
As of this commit, the SFrameTest passes regardless of the placement
of an invisible frame. This commit does not seem to fix all cases for
the E-frame, so this will still need to be investigated in a followup.

Change-Id: I631e29031afa427bfd54ff0e4793665f85c1bd1c