Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (1 comment)

I found a reference of this feature[1]. This paper described how the hash function is designed.

"The first-level hash value is first used to find the candidates approximated to the current block. If more than one block is found, the best prediction block is chosen from
the candidates by comparing the second-level hash values."

Two hash functions are adopted. 
* The first hash function used is designed to find the prediction blocks approximated to the target block. However traditional hash functions like CRC can only be used to find blocks identical to the input block. As a result, traditional hash functions do not help to find blocks which approximates to the input block. This paper proposed another hash function which is not a traditional hash function like CRC.
* The second hash function is a CRC function.

But in the code of AV1, it seems does not follow the design, both hash function is the same CRC just with different POLY. Is this also the desired behavior?

[1]: W. Zhu, W. Ding, J. Xu, Y., Shi, B., Yin, "Hash-based block matching for screen content coding", IEEE Trans. Multimedia, vol. 17, no. 7, pp. 935-944, Jul. 2015.

Patch-set: 2
Reviewer: Gerrit User 8900 <8900@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>