Update patch set 3
Patch Set 3:

With this approach, it does not seem to be possible to generate single-frame video-type sequences. This can be useful for many reasons: conformance testing, comparison video/still ... Would be possible to have a command-line option (--reduced-header) to use the still, reduced header when both are true: --limit=1 (or single frame input) and --reduced-header=1?

Patch-set: 3
CC: Gerrit User 8014 <8014@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>