[NORMATIVE] Drop error-resilient bit for key frames
For key frames, the error-resilient mode bit is not
required. So drop it.

Note for intra only frames, error-resilient mode can
still be used to set the order hints for all the reference
buffers explicitly, since not all buffers are refreshed.
But for key frames since all buffers are refreshed, there is
no need to send the bit.

BUG=aomedia:1741

Change-Id: I31c160fb01d90c2799e30551b7f757f57d98409e