[NORMATIVE] Code coded_segid with a limited range
This commit is a normative change to allow for more efficient coding
of segment IDs. Previously the code assumed that all coded segment
ID values are permitted, including in the range where they've been
explicitly signalled as non-active. But in reality, the coded values
for segment IDs will never exceed the last active segment ID, as the
neg_(de)interleave functions will error out.
Hence, to improved coding if the number of segment IDs used is below
the maximum, code them within the range guaranteed.

For a 30 frame 240x170 video there's an overall rate decrease of 0.04%.
The gains will be greater for larger videos and number of frames.

Change-Id: I80e804a2d626a85df6676b07499c77e13dee1bd7