Optimize loop restoration filter
While there is no cdef and no superres, we can write an optimized loop
restoration filter. This patch removes the operations to save boundary
lines for each 64x64 processing unit. During setup and restore of the
stripe boundary, we only need to save and restore 1 line from above
and 1 line from below. The boundaries buffer is not used.
(TODO: not allocate memory for boundaries buffer in this case.)

This change doesn't cause bitstream change. Test results verified that.

Change-Id: I26588154fd8de79c094f91a86a512586c10e1876