[NORMATIVE] Rearrange error-resilience bit in hdr
This patch backs out the change in:
https://aomedia-review.googlesource.com/c/aom/+/55215
and instead changes the conditions under which order
hints will be sent by rearranging the hdr so that the
refresh frame flags are available prior to deciding
whether to send the explicit order hints.

Now as before, the error-resilient bit is sent for all
non-S frames including Key frames, but it will simply
be useless for cases like:
KEY_FRAME with refresh_flags = 0xFF.
A further header rearrangement will be needed to drop the
error resilient bit for the above case, but there is an
advantage in keeping the error_resilient bit available
near the top of the header for easier parsing. So I am not
sure if it is worth doing.

BUG=aomedia:1741

Change-Id: Ifda36fdab7bdcfd031cff62da60efd33daec3334