Update patch set 3
Patch Set 3:

This change makes me nervous because the adjustment to the MV cost where no motion field is available was put in specifically to help with a severe visual artifact (for example irregular motion field in clouds) in the full knowledge that it would likely not help or even hurt metrics a little.

When my desktop data is retrieved I will try and dig out specific samples or test clips to verify whether this is still needed in AOM.

If it is shown that this can be safely deleted the #define NEW_MV_DISCOUNT_FACTOR should also go as this is only used in this one place.

Patch-set: 3
Reviewer: Gerrit User 5610 <5610@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>