Refactor OBU header+size parsing code
av1_decode_frame_from_obus() currently has a block of code to handle
OBU headers and to correctly derive the payload size in both the
annexb and non-annexb cases.

This patch moves that code into its own function, so that it can be
shared with other places which do OBU parsing (in particular,
decoder_peek_si_internal()).

Along the way, tighten up the error detection, so that we're more
likely to detect cases where the input buffer is too short.

BUG=aomedia:1794

Change-Id: If16fe25b3a044743d15f44b9c91c2f3ce7c49df2