obudec: Allow OBU-format still-image streams
Refactor the file_is_obu() function so that it allows files
beginning with an OBU_SEQUENCE_HEADER. This may occur for still-image
streams, where we do not require an initial OBU_TEMPORAL_DELIMITER.

In order to avoid needing a large detection buffer, we factor out
the header + payload size parsing from obudec_read_one_obu(), so that
we can detect an OBU_SEQUENCE_HEADER without having to read its payload.
Then, if the file is determined to be in OBU format, we can read the
payload at that point.

BUG=aomedia:1805

Change-Id: Id7a5cbc9388dd0dce9ba94e78a16c4c8456f6755