Clean up GF structure
The use of the mid-altref is indeed a concept for the libvpx encoder.
Such mid-altrefs have never been used in libaom, as it has adopted
the use of a hierarchical GF structure. This CL is to remove the code
used only in libvpx and hence have the GF structure implementation in
libaom unified.

Change-Id: I6a4526eaa21541a8e3bf971a74079292dde5383c