Dynamically-sized buffer in obudec.c
The buffer used to read data from .obu files is set to be
OBU_BUFFER_SIZE = 500 KiB. Because of this, prior to this patch
a temporal unit of > 500KiB resulted in an error (when reading a
non-annex_b stream), causing decoding to halt.

This patch dynamically increases the size of the buffer as it is
needed.

BUG=aomedia:1812

Change-Id: Ie2f4a85efea37d37b9d41f182cdf31913a82e5cb