Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Does this change have any normative effect?
> 
> In other words, does the output of the decoder for valid streams depend on whether this patch is included?

This change is normative and seems not to do what you expect.  In read_sequence_header_obu() the sequence is:
1). read some of the seq params - ie still_picture, reduced_still_picture, operating_point_idc[], level[], decoder_rate_model_param_present_flag[]
2). read_sequence_header()

This change has the effect that #2 sets all the values read in #1 to 0.  I'll raise a bug and I suggest we revert this patch for now.

Also what is buganizer?  How do we see what the intention of this change is?

Patch-set: 2
CC: Gerrit User 6145 <6145@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>