Fix sse overflow in highbd_*_obmc_variance sse4 version
Also the final sse will be stored as an unsigned 64-bit integer,
which is enough. The cumulation process is done in 4 x int32.
Therefore, for 10-bit and 12-bit video, int32 is not enough to
hold 1/4 of the total sse of large blocks.
In this patch, max_pel_allowed_per_ovf is set up for 10-bit and
12-bit cases, hbd_obmc_variance_w8n is applied to patches sizes at
                   bw * (max_allowed_per_ovf / bw)
to avoid overflow.

BUG=aomedia:1772

Change-Id: Iaae2e1403baa93c4ba454d6760281f95428f437a