Restrict number of workers in the encoder
Prior to this patch, the encoder:

- initialised a number of workers based on the width of the *first*
  frame; and

- decided how many workers to make use of based on the width of the
  *current* frame.

This resulted in problems when using superres, as later frames could be
larger than the first frame, causing the encoder to attempt to access
workers that did not exist.

This patch ensures that the number of workers the encoder decides to use
is capped by the number of workers initialised on the first frame.

I have also updated the horz superres test to test tile-columns = 4, as
this no longer fails.

BUG=aomedia:1494

Change-Id: I022c7e9843464e590c4b72e2a7bf47ef8afa1b0b