Fix the source address of memmove in obudec_read_temporal_unit().
This bug was introduced in commit 4beef21b7d4fedac0247683e988f6b8106029395.
When we break out of the while (1) loop (after reading a temporal unit),
obu_ctx->bytes_buffered is not incremented. Therefore,
obu_ctx->buffer + obu_ctx->bytes_buffered points to the beginning of
the OBU that was just read into the buffer.

Change-Id: I13e41c1e0a0387f90536882756665849c3ce02f6