Avoid intrinsics incompatible with VS compiler
TESTING ONLY: This patch forces the workarounds to be used so that they
can be tested.

The intrinsics _mm_set_epi64x(), _mm_set1_epi64x(), _mm256_set_epi64x()
and _mm256_set1_epi64x() are incompatible with Visual Studio x86
configurations. This patch replaces calls to these intrinsics with calls
to functions that use a workaround when in such a configuration.

In addition, a small amount of tidying up has been done to the SIMD code.

BUG=aomedia:1815

Change-Id: Ic20f64c51bc4a6d12c18c18a69f492da33b72a0f