Update patch set 3
Patch Set 3:

(2 comments)

I was thinking that above_context and left_context should be handled in the same way, since now we support both row and column tiles. It doesn't matter which way to use. If this is too complicated, we can leave it as a "TODO" for now. Thanks.

Patch-set: 3
Reviewer: Gerrit User 5185 <5185@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>