Update patch set 7
Patch Set 7:

(2 comments)

> Patch Set 7:
> 
> (8 comments)
> 
> > Patch Set 6:
> > 
> > (6 comments)
> > 
> > It looks good overall. Thank you for getting it done. I only have some questions. This should be covered by multi-thread unit test, right? Have you done more testing to make sure it generates the same bit stream as using single thread? Also, did you test the MT decoder speed with your change?
> 
> The decoder bit-exactness is tested by the test 'AV1DecodeMultiThreadedTest' and the encoder bit-exactness by 'AVxEncoderThreadTest'.

We are seeing a decode time reduction of 23% for 1080p sequence and 13% for 720p sequence with 4 threads due to loop-filter multi-threading (Please note that decoder multi-threading is enabled both in reference and current version). The scaling achieved by loop-filtering for 4 threads (1080p) is 3.3. The deficient of 0.7 is due to the processor clock drop and multi-threading dependencies/overheads.

Patch-set: 7
Reviewer: Gerrit User 8893 <8893@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>